moving our work into mcloud.go file, adding UnmarshallParam function
